Query Flickr API and build a dataframe of metadata

This function uses the Flickr API to get metadata about the images

photosSearch(year_range, text, bbox = NULL, woe_id = NULL, has_geo = TRUE)

Arguments

year_range
The year to search between (inclusive). For example c(2005,2007) will search for images from 2005, 2006 and 2007

text
Set keywords so that query returns pictures with the text (e.g. "bird")

bbox
A vector of 4 values defining the Bounding Box of the area that will be searched. The 4 values represent the bottom-left corner of the box and the top-right corner, minimum_longitude, minimum_latitude, maximum_longitude, maximum_latitude.

woe_id
A 32-bit identifier that uniquely represents spatial entities (e.g. 12578048 is Scotland). These ids can be found using the findPlaces function Can not be provided if bbox allready provided

has_geo
Logical, should the function only return records that have spatial data.

Value

A dataframe of metadata

See also

findPlaces

Examples

# run a workflow, using the logistic regression model

# NOT RUN {
# Authenticate the first time you run a search
#auth.flickr()

birds <- photosSearch(year_range = c(2005,2006),
                      text = "mute+swan",
                      woe_id =  12578048,
                      has_geo = TRUE)

head(birds)

# }